moved enemy related things to appropriate map
