+++ ContextDesktop/src/main/java/com/github/migi_1/Context/enemy/EnemyGenerator.java
@@ -1,4 +1,4 @@
-package com.github.migi_1.Context.model.Enemy;
+package com.github.migi_1.Context.enemy;
 
 import java.util.LinkedList;
